# nsqd: optimize guid generation