@mreiferson 2015-09-30
@ploxiln any of this worth landing?  I'm a little concerned about micro-optimizations like bit-shift-for-division, but maybe that one?