@ploxiln 2015-09-30
I'm not sure this should be merged as-is, I expect we'll want to tweak it. We may not need the subjective uglyness of the additional argument to NewGUID() - it seems that errors shouldn't be that common ...

I like the shift optimization - it makes sense that it could be noticeably helpful on ARM, I'll check.