@mreiferson 2015-09-30
Let's drop the time stuff and leave the shift optimization and land it.