@ploxiln 2015-10-01
I've reduced this PR to just the shift optimization. But I haven't tested it on arm yet, and I don't think there's any rush.

I'm in the process of getting stuff set up on a raspberrypi 2 which has 4 Cortex A7 cores, which are lower-power and in-order rather than out-of-order but from the same generation as the Cortex A15. And it's also an interesting target in any case.