@ploxiln 2015-10-01
It's probably worth mentioning that the time-part of the id will jump back 18577 hours. I don't think there's much chance of overlap; that's about as long as nsq has existed.

```
func main() {
	ns := time.Now().UnixNano()
	ms := ns / 1e6
	pms := ns >> 20
	fmt.Printf("milliseconds = %d\n", ms);
	fmt.Printf("pseudo-milliseconds = %d\n", pms);
	fmt.Printf("difference = %d\n", ms - pms);
	fmt.Printf("hours = %d\n", (ms - pms) / (1000 * 3600));
}
[pierce@plo-pro current]$ go run test.go
milliseconds = 1443665428226
pseudo-milliseconds = 1376786640383
difference = 66878787843
hours = 18577
```