@ploxiln 2015-10-01
Did have a noticeable improvement for pub:

```
Switched to branch 'master'
# using --mem-queue-size=1000000 --data-path= --size=200 --batch-size=200
PUB: [bench_writer] 2015/10/01 01:25:05 duration: 10.004711496s - 5.838mb/s - 30605.580ops/s - 32.674us/op
SUB: [bench_reader] 2015/10/01 01:25:15 duration: 10.016315311s - 1.698mb/s - 8900.379ops/s - 112.355us/op
Switched to branch 'time_div_optim'
# using --mem-queue-size=1000000 --data-path= --size=200 --batch-size=200
PUB: [bench_writer] 2015/10/01 01:26:25 duration: 10.004926446s - 6.470mb/s - 33923.288ops/s - 29.478us/op
SUB: [bench_reader] 2015/10/01 01:26:35 duration: 10.019001199s - 1.698mb/s - 8904.979ops/s - 112.297us/op
Switched to branch 'master'
# using --mem-queue-size=1000000 --data-path= --size=200 --batch-size=200
PUB: [bench_writer] 2015/10/01 01:27:31 duration: 10.004017781s - 5.941mb/s - 31147.486ops/s - 32.105us/op
SUB: [bench_reader] 2015/10/01 01:27:41 duration: 10.025512223s - 1.694mb/s - 8879.945ops/s - 112.613us/op
Switched to branch 'time_div_optim'
# using --mem-queue-size=1000000 --data-path= --size=200 --batch-size=200
PUB: [bench_writer] 2015/10/01 01:28:36 duration: 10.004349052s - 6.558mb/s - 34385.046ops/s - 29.082us/op
SUB: [bench_reader] 2015/10/01 01:28:46 duration: 10.015671236s - 1.712mb/s - 8977.831ops/s - 111.385us/op
```

Also: many ARM SoC have multiple cores, and GOMAXPROCS=2 *almost* eliminates the difference this branch makes, for this simple bench.sh (pretty consistent pub perf of 10.5 mb/s -> 10.8 mb/s)