@ploxiln 2015-10-01
There were a few options for twepoch, I decided to keep it almost exactly the same but update the comment.

I've done as much half-hearted ARMv7 benchmarking as I'm inclined to do. fwiw