@mreiferson 2015-10-01
@ploxiln correct me if I'm wrong, but won't this also change the number of unique IDs we're able to generate per second?