@ploxiln 2015-10-01
Yes, there will be 5% fewer unique IDs per second.

```python
>>> (10**9 / 2**20) * 1.0 / 1e3 
0.953
```