@ploxiln 2015-10-01
So, it reduces the theoretical max messages per workerID per second from 4,096,000 to 3,903,488